Adiciona página de cadastro de consultas com formulário e lógica de manipulação
